ml/requirements.txt
# ML service dependencies
flask>=2.3.0
flask-cors>=4.0.0
langchain>=0.0.200
langchain-community>=0.0.10
chromadb>=0.4.18
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uuid>=1.30

# YouTube transcription dependencies
yt-dlp>=2023.3.4
openai-whisper>=20230314
ffmpeg-python>=0.2.0
